move sticky logic to header
